Fix SphPolygon producing unexpected results for 32-bit float coordinates (#385)
* Fix geometry objects' bbox not being clockwise

* Force geometry get_bbox_lonlats to always return numpy arrays

* Remove TODO in new test that cna happen later

* Rename variable to avoid redefining builtin sum

* Remove type annotations that aren't allowed in Python >3.7

* Fix SphPolygon failing for specific 32-bit float coordinates

+++ pyresample/spherical.py
@@ -330,7 +330,7 @@ class SphPolygon(object):
     """
 
     def __init__(self, vertices, radius=1):
-        self.vertices = vertices
+        self.vertices = vertices.astype(np.float64, copy=False)
         self.lon = self.vertices[:, 0]
         self.lat = self.vertices[:, 1]
         self.radius = radius
